Refactoring Agent and FMUAgent to save matches and heuristics by level
